+++ src/pathsampler.hpp
@@ -5,17 +5,17 @@
 
 class PathSampler {
 public:
-	PathSampler(size_t total_number);
+	PathSampler(unsigned short total_number);
 	// select one random subset of paths
-	void select_single_subset(std::vector<size_t>& result, size_t sample_size) const;
+	void select_single_subset(std::vector<unsigned short>& result, unsigned short sample_size) const;
 	// select random subset of paths n times
-	void select_multiple_subsets(std::vector<std::vector<size_t>>& result, size_t sample_size, size_t n) const;
+	void select_multiple_subsets(std::vector<std::vector<unsigned short>>& result, unsigned short sample_size, unsigned short n) const;
 	// partition set of paths
-	void partition_paths(std::vector<std::vector<size_t>>& result, size_t sample_size) const;
+	void partition_paths(std::vector<std::vector<unsigned short>>& result, unsigned short sample_size) const;
 	// partition paths in sample aware manner
-	void partition_samples(std::vector<std::vector<size_t>>& result, size_t sample_size) const;
+	void partition_samples(std::vector<std::vector<unsigned short>>& result, unsigned short sample_size) const;
 private:
-	size_t total_number; 
+	unsigned short total_number; 
 };
