+++ src/transitionprobabilitycomputer.hpp
@@ -9,8 +9,8 @@
 
 class TransitionProbabilityComputer {
 public:
-	TransitionProbabilityComputer(size_t from_variant, size_t to_variant, double recomb_rate, size_t nr_paths, bool uniform = false, long double effective_N = 25000.0L);
-	long double compute_transition_prob(size_t path_id1, size_t path_id2, size_t path_id3, size_t path_id4);
+	TransitionProbabilityComputer(size_t from_variant, size_t to_variant, double recomb_rate, unsigned short nr_paths, bool uniform = false, long double effective_N = 25000.0L);
+	long double compute_transition_prob(unsigned short path_id1, unsigned short path_id2, unsigned short path_id3, unsigned short path_id4);
 private:
 	std::vector<long double> probabilities;
 	bool uniform;
